requirements.txt
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.35
langgraph==0.2.16
langchain-groq==0.1.3

faiss-cpu==1.8.0
sentence-transformers==2.7.0

python-dotenv==1.0.1
scikit-learn==1.5.1